refactor#74 social login api
