+++ data/oauth/src/main/java/com/teamsolply/solply/oauth/dto/request/SocialLoginRequestDto.kt
@@ -1,0 +1,10 @@
+package com.teamsolply.solply.oauth.dto.request
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SocialLoginRequestDto(
+    @SerialName("oauthAccessToken")
+    val oauthAccessToken: String,
+)
