+++ remote/oauth/src/main/java/com/teamsolply/solply/oauth/service/OauthService.kt
@@ -1,15 +1,16 @@
 package com.teamsolply.solply.oauth.service
 
 import com.teamsolply.solply.network.model.BaseResponse
+import com.teamsolply.solply.oauth.dto.request.SocialLoginRequestDto
 import com.teamsolply.solply.oauth.dto.response.SocialLoginResponseDto
 import retrofit2.http.Body
 import retrofit2.http.POST
 import retrofit2.http.Path
 
 interface OauthService {
-    @POST("api/auth/social/{provider}/login")
+    @POST("api/auth/social/{soicialPlatform}/login")
     suspend fun postSocialLogin(
-        @Path("provider") provider: String,
-        @Body oauthAccessToken: String
+        @Path("soicialPlatform") soicialPlatform: String,
+        @Body oauthAccessToken: SocialLoginRequestDto
     ): BaseResponse<SocialLoginResponseDto>
 }
